Edited gitignore

+++ bot.py
@@ -22,6 +22,9 @@ intents.message_content = True
 
 client = discord.Client(intents=intents)
 
+with open('whitelist.txt', 'r') as file:
+    whitelist = [line.strip() for line in file.readlines()]
+
 successfulJobs = 0
 failedJobs = 0
 consecutiveFailedJobs = 0
@@ -48,17 +51,23 @@ async def on_reaction_add(reaction, user):
         if reaction.me:
             # await user.send("You reacted with the same emoji as the bot! This feature is coming soon") # this is for DMing the person who reacted
             # await reaction.message.channel.send(f"{user} reacted with the same emoji as the bot! This feature is coming soon")
-            print(f'{user} reacted on message with content {user_message} in #{channel} in guild {guild}. Starting new job.')
-            try:
-                tryToSendMessage = await reaction.message.channel.send(f'Attempting to start job...')
-            except Exception as ex:
-                await user.send("It appears that I may not have permission to send the video in the channel. Here's more info on what went wrong:")
-                template = "||{0}||"
-                errorToSend = template.format(ex)
-                await user.send(errorToSend)
+
+            # Check if player has permissions to use bot
+            if str(user.id) in whitelist:
+                print(f'{user} reacted on message with content {user_message} in #{channel} in guild {guild}. Starting new job.')
+                try:
+                    tryToSendMessage = await reaction.message.channel.send(f'Attempting to start job...')
+                except Exception as ex:
+                    await user.send("It appears that I may not have permission to send the video in the channel. Here's more info on what went wrong:")
+                    template = "||{0}||"
+                    errorToSend = template.format(ex)
+                    await user.send(errorToSend)
+                else:
+                    await tryToSendMessage.delete()
+                    await CreatePreview(reaction.message, username, user_message)
             else:
-                await tryToSendMessage.delete()
-                await CreatePreview(reaction.message, username, user_message)
+                await user.send("Hey there, you'll need to purchase the Instagram reel previewer module for me to post Instagram reels on Discord. Get it here: ")
+                print(f'{user} does not have module')
 
 @client.event
 async def on_message(message):
@@ -81,7 +90,12 @@ async def on_message(message):
             print("Putting a download reaction emoji")
             await message.add_reaction("⏬")
         else:
-            await CreatePreview(message, username, user_message)
+            if str(message.author.id) in whitelist:
+                await CreatePreview(message, username, user_message)
+            else:
+                #await message.author.send("Hey there, you'll need to purchase the Instagram reel previewer module for me to post Instagram reels on Discord. Get it here: ")
+                await message.channel.send("Hey there, you'll need to purchase the Instagram reel previewer module for me to post Instagram reels on Discord. Get it here: ")
+                print(f'{message.author} does not have module')
 
 async def CreatePreview(message, username, user_message):
     timeElapsed = time.time()
